Really test that caseService sends notification messages to the broker
the spring-cloud-stream-test-support has been deprecated and doesn't seem to work
for producers (only works for functions ?), so use the newer spring integration test binder.

The previous test was sending the message itself instead of testing that
caseService was doing it. To do it in the CaseControllerTest, we need a more
complete spring context so we switch to @SpringBootTest .

Signed-off-by: Jon Harper <jon.harper87@gmail.com>